+++ lib/notesclub/x/x_token.ex
@@ -1,15 +1,14 @@
-defmodule Notesclub.X.XToken do
+defmodule Notesclub.X.XTokens.XToken do
   @moduledoc """
   Schema for storing X API access tokens
   """
-  use Ecto.Schema
+  use TypedEctoSchema
   import Ecto.Changeset
-  import Ecto.Query, warn: false
 
   alias Notesclub.Repo
   alias Notesclub.X.XToken
 
-  schema "x_tokens" do
+  typed_schema "x_tokens" do
     field :access_token, :string
     field :refresh_token, :string, default: nil
     field :last_used_at, :utc_datetime
@@ -22,40 +21,5 @@ defmodule Notesclub.X.XToken do
     token
     |> cast(attrs, [:access_token, :refresh_token, :last_used_at])
     |> validate_required([:access_token, :last_used_at])
-  end
-
-  @doc """
-  Returns the most recent X API token.
-  """
-  def get_latest_token do
-    XToken
-    |> last(:inserted_at)
-    |> Repo.one()
-  end
-
-  @doc """
-  Creates a new X API token.
-  """
-  def create_token(attrs \\ %{}) do
-    attrs = Map.put(attrs, :last_used_at, DateTime.utc_now())
-
-    %XToken{}
-    |> changeset(attrs)
-    |> Repo.insert()
-  end
-
-  def update_token(token, attrs) do
-    token
-    |> changeset(attrs)
-    |> Repo.update()
-  end
-
-  @doc """
-  Updates the last_used_at field for a token.
-  """
-  def mark_token_used(token) do
-    token
-    |> changeset(%{last_used_at: DateTime.utc_now()})
-    |> Repo.update()
   end
 end
